Added Departments Service and Controller

+++ demo/src/main/java/com/example/demo/repositories/DepartmentsRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartmentsRepository extends CrudRepository<Departments, Long> {}
+public interface DepartmentsRepository extends CrudRepository<Departments, Integer> {}
